[BUGFIX] Suggest wizard does not work on new records
The suggest wizard gets it's parameters via a POST json
body. Inside there is the current record uid given,
which follows the scheme "NEW{random number}". For whatever
reason the wizard casts this uid to an integer which results
in a 0. The following check for a numeric uid fails in the
case of a new record, as it is always numeric because of the
cast.

This patch removes the cast. No further sanitizing is necessary
as the uid is casted in every further use case where appropriate.

Resolves: #71056
Releases: master

Reviewed-on: https://review.typo3.org/44412
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Gernot Schulmeister <gernotschulmeister@gmx.at>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>

+++ typo3/sysext/backend/Classes/Form/Wizard/SuggestWizard.php
@@ -142,7 +142,7 @@ class SuggestWizard
         $search = isset($parsedBody['value']) ? $parsedBody['value'] : $queryParams['value'];
         $table = isset($parsedBody['table']) ? $parsedBody['table'] : $queryParams['table'];
         $field = isset($parsedBody['field']) ? $parsedBody['field'] : $queryParams['field'];
-        $uid = (int)(isset($parsedBody['uid']) ? $parsedBody['uid'] : $queryParams['uid']);
+        $uid = isset($parsedBody['uid']) ? $parsedBody['uid'] : $queryParams['uid'];
         $pageId = (int)(isset($parsedBody['pid']) ? $parsedBody['pid'] : $queryParams['pid']);
         $newRecordRow = isset($parsedBody['newRecordRow']) ? $parsedBody['newRecordRow'] : $queryParams['newRecordRow'];
         // If the $uid is numeric, we have an already existing element, so get the
